backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
httpx>=0.26.0
aiofiles>=23.2.1
tencentcloud-sdk-python>=3.0.1000
cos-python-sdk-v5>=1.9.30
beautifulsoup4>=4.12.3
redis>=5.0.1
# Vector Search / RAG
pgvector>=0.2.4
openai>=1.12.0
tiktoken>=0.5.2
sse-starlette>=2.0.0
markdown2>=2.4.12
weasyprint>=62.3
